use int for activity name; run script adjusted; file name in counts script

+++ src/main/java/org/matsim/prepare/CreatePopulation.java
@@ -588,7 +588,7 @@ public class CreatePopulation {
 					durationCategoryNr = 1;
 				}
 				
-				String newType = act.getType() + "_" + (durationCategoryNr * timeCategorySize);
+				String newType = act.getType() + "_" + (int) (durationCategoryNr * timeCategorySize);
 				act.setType(newType);				
 			}
 		}
